requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & LLM
langchain==0.1.4
langchain-community==0.0.17
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.10.0
anthropic==0.18.1

# RAG & Vector Store
chromadb==0.4.22
sentence-transformers==2.3.1
faiss-cpu==1.7.4
llama-index==0.10.0

# Multi-modal Processing
Pillow==10.2.0
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
transformers==4.37.2
openai-whisper==20231117

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
